[IMP] account: open_move_widget owl migration
