+++ addons/account/static/src/js/open_move_widget.js
@@ -1,23 +1,21 @@
 /** @odoo-module **/
-import fieldRegistry from 'web.field_registry';
-import { FieldChar } from 'web.basic_fields';
 
-const OpenMoveWidget = FieldChar.extend({
-    template: 'account.OpenMoveTemplate',
-    events: Object.assign({}, FieldChar.prototype.events, {
-        'click': '_onOpenMove',
-    }),
-    _onOpenMove: function(ev) {
-        ev.preventDefault();
-        ev.stopPropagation();
-        var self = this;
-        this._rpc({
-            model: 'account.move.line',
-            method: 'open_move',
-            args: [this.res_id],
-        }).then(function (actionData){
-            return self.do_action(actionData);
-        });
-    },
-});
-fieldRegistry.add('open_move_widget', OpenMoveWidget);
+import { registry } from "@web/core/registry";
+import { useService } from "@web/core/utils/hooks";
+import { CharField } from "@web/views/fields/char/char_field";
+
+class OpenMoveWidget extends CharField{
+    setup() {
+        super.setup();
+        this.orm = useService("orm");
+        this.action = useService("action");
+    }
+
+    async _onOpenMove(ev) {
+        const act = await this.orm.call("account.move.line", "open_move", [this.props.record.resId], {})
+        this.action.doAction(act);
+    }
+}
+
+OpenMoveWidget.template = "account.OpenMoveWidget";
+registry.category("fields").add("open_move_widget", OpenMoveWidget);
