+++ addons/account/static/src/xml/open_move_widget.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<templates>
+
+    <t t-name="account.OpenMoveWidget" owl="1">
+        <a href="#" t-esc="props.value" t-on-click.prevent="(ev) => this._onOpenMove()"/>
+    </t>
+
+</templates>
